Move util functions into a separate file.
